+++ tests/Provision2016Full.ps1
@@ -126,7 +126,7 @@ try {
         -LicenseKey WCPQN-33442-VH2RQ-M4RKF-GXYH4 `
         -InstallDir "c:\Program Files\Microsoft Dynamics CRM" `
         -CreateDatabase `
-        -SqlServer $dbHostName\SPIntra01 `
+        -SqlServer $dbHostName\SQLInstance01 `
         -PrivUserGroup "CN=CRM01PrivUserGroup,OU=CRM groups,DC=contoso,DC=local" `
         -SQLAccessGroup "CN=CRM01SQLAccessGroup,OU=CRM groups,DC=contoso,DC=local" `
         -UserGroup "CN=CRM01UserGroup,OU=CRM groups,DC=contoso,DC=local" `
@@ -148,7 +148,7 @@ try {
         -BaseCurrencySymbol `$ `
         -BaseCurrencyPrecision 2 `
         -OrganizationCollation Danish_Norwegian_CI_AI `
-        -ReportingUrl http://$dbHostName/ReportServer_SPIntra01 `
+        -ReportingUrl http://$dbHostName/ReportServer_RSInstance01 `
         -InstallAccount $CRMInstallAccountCredential
 } catch {
     Write-Host $_.Exception.Message -ForegroundColor Red;
@@ -180,13 +180,13 @@ try {
     if ( $dbHostName -eq $env:COMPUTERNAME ) {
         Install-Dynamics365ReportingExtensions `
             -MediaDir C:\Install\Dynamics\CRM2016RTMDan\SrsDataConnector `
-            -InstanceName SPIntra01 `
+            -InstanceName SQLInstance01 `
             -InstallAccount $CRMInstallAccountCredential
     } else {
         Install-Dynamics365ReportingExtensions `
             -MediaDir \\$env:COMPUTERNAME\c$\Install\Dynamics\CRM2016RTMDan\SrsDataConnector `
             -ConfigDBServer $dbHostName `
-            -InstanceName SPIntra01 `
+            -InstanceName SQLInstance01 `
             -InstallAccount $CRMInstallAccountCredential
     }
 } catch {