+++ tests/Provision90Full.ps1
@@ -122,7 +122,7 @@ try {
         -LicenseKey KKNV2-4YYK8-D8HWD-GDRMW-29YTW `
         -InstallDir "c:\Program Files\Microsoft Dynamics CRM" `
         -CreateDatabase `
-        -SqlServer $dbHostName\SPIntra01 `
+        -SqlServer $dbHostName\SQLInstance01 `
         -PrivUserGroup "CN=CRM01PrivUserGroup,OU=CRM groups,DC=contoso,DC=local" `
         -SQLAccessGroup "CN=CRM01SQLAccessGroup,OU=CRM groups,DC=contoso,DC=local" `
         -UserGroup "CN=CRM01UserGroup,OU=CRM groups,DC=contoso,DC=local" `
@@ -144,7 +144,7 @@ try {
         -BaseCurrencySymbol `$ `
         -BaseCurrencyPrecision 2 `
         -OrganizationCollation French_CI_AI `
-        -ReportingUrl http://$dbHostName/ReportServer_SPIntra01 `
+        -ReportingUrl http://$dbHostName/ReportServer_RSInstance01 `
         -InstallAccount $CRMInstallAccountCredential
 } catch {
     Write-Host $_.Exception.Message -ForegroundColor Red;
@@ -177,13 +177,13 @@ try {
     if ( $dbHostName -eq $env:COMPUTERNAME ) {
         Install-Dynamics365ReportingExtensions `
             -MediaDir C:\Install\Dynamics\Dynamics365Server90RTMFra\SrsDataConnector `
-            -InstanceName SPIntra01 `
+            -InstanceName SQLInstance01 `
             -InstallAccount $CRMInstallAccountCredential
     } else {
         Install-Dynamics365ReportingExtensions `
             -MediaDir \\$env:COMPUTERNAME\c$\Install\Dynamics\Dynamics365Server90RTMFra\SrsDataConnector `
             -ConfigDBServer $dbHostName `
-            -InstanceName SPIntra01 `
+            -InstanceName SQLInstance01 `
             -InstallAccount $CRMInstallAccountCredential
     }
 } catch {